only rely on OS APIs when necessary
