+++ Cargo.toml
@@ -15,11 +15,13 @@ categories = ["no-std"]
 [badges]
 maintenance = { status = "actively-developed" }
 
-[target.'cfg(any(target_os = "linux", target_os = "android"))'.dependencies]
+# Only pull in OS libraries on platforms where we don't have an ASM implementation
+
+[target.'cfg(all(not(any(target_arch="x86",target_arch="x86_64",target_arch="loongarch64")),any(target_os="linux", target_os="android")))'.dependencies]
 libc = { version = "0.2", default-features = false }
 
-[target.'cfg(target_os = "windows")'.dependencies]
-windows = { version = "0.61", features = [
+[target.'cfg(all(not(any(target_arch="x86",target_arch="x86_64",target_arch="loongarch64")),target_os="windows"))'.dependencies]
+windows-sys = { version = "0.61", features = [
   "Win32_System_Threading",
   "Win32_System_Diagnostics_Debug",
 ] }
