+++ src/lib.rs
@@ -1,20 +1,23 @@
 #![doc = include_str!("../README.md")]
 #![no_std]
 
-#[cfg(any(target_os = "linux", target_os = "android"))]
-mod linux;
-#[cfg(any(target_os = "linux", target_os = "android"))]
-use linux::*;
-
-#[cfg(target_os = "windows")]
-mod windows;
-#[cfg(target_os = "windows")]
-use windows::*;
-
-#[cfg(target_os = "macos")]
-mod macos;
-#[cfg(target_os = "macos")]
-use macos::*;
+#[cfg_attr(
+    any(
+        target_arch = "x86",
+        target_arch = "x86_64",
+        target_arch = "loongarch64",
+    ),
+    path = "arch.rs"
+)]
+#[cfg_attr(
+    not(any(
+        target_arch = "x86",
+        target_arch = "x86_64",
+        target_arch = "loongarch64",
+    )),
+    path = "os.rs"
+)]
+mod clear_cache_impl;
 
 /// Flush CPU's instruction cache at given range.
 ///
@@ -26,5 +29,5 @@ use macos::*;
 /// instructions and syscalls make it difficult to guarantee that this function is totally
 /// safe.
 pub unsafe fn clear_cache<T>(start: *const T, end: *const T) -> bool {
-    os_arch_clear_cache(start, end)
+    clear_cache_impl::os_arch_clear_cache(start, end)
 }
